--- src/lib/axios/requests/get-products.examples.ts
@@ -1,1349 +1,0 @@
-const data = {
-  count: 7,
-  limit: 7,
-  offset: 0,
-  results: [
-    {
-      categories: [
-        {
-          id: 'b86a46f7-70a9-483f-b74b-a7b4ac95ddda',
-          typeId: 'category',
-        },
-      ],
-      categoryOrderHints: {},
-      createdAt: '2024-05-23T05:54:46.722Z',
-      description: {
-        'en-US': 'A timeless bouquet featuring a dozen fresh red roses, elegantly arranged with greenery.',
-      },
-      hasStagedChanges: false,
-      id: '5a2b875e-68f9-42c2-83f7-67f5418a55ca',
-      key: 'classic_rose_bouquet',
-      lastModifiedAt: '2024-06-04T13:20:15.329Z',
-      masterVariant: {
-        attributes: [
-          {
-            name: 'color',
-            value: ['green', 'red'],
-          },
-          {
-            name: 'size',
-            value: 'medium',
-          },
-        ],
-        id: 1,
-        images: [
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/classic_rose_bouquet-oBl_4aEu.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/classic_rose_bouquet-gBL67E9D.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/classic_rose_bouquet-SVfPmFL7.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/classic_rose_bouquet-N496xh6R.jpg',
-          },
-        ],
-        prices: [
-          {
-            id: '1df0bacc-adf1-4d3f-9dbd-8882def9f079',
-            key: 'classic_rose_bouquet_standard_price',
-            validFrom: '2024-05-30T21:00:00.000Z',
-            validUntil: '2024-06-29T21:00:00.000Z',
-            value: {
-              centAmount: 2999,
-              currencyCode: 'EUR',
-              fractionDigits: 2,
-              type: 'centPrecision',
-            },
-          },
-        ],
-      },
-      metaDescription: {},
-      metaTitle: {},
-      name: {
-        en: 'Classic Rose Bouquet',
-        'en-US': 'Classic Rose Bouquet',
-      },
-      priceMode: 'Embedded',
-      productType: {
-        id: '8ca55f27-b8ef-4321-aa32-7313517e00ba',
-        typeId: 'product-type',
-      },
-      published: true,
-      searchKeywords: {},
-      slug: {
-        'en-US': 'classic-rose-bouquet',
-      },
-      variants: [],
-      version: 25,
-    },
-    {
-      categories: [
-        {
-          id: 'b86a46f7-70a9-483f-b74b-a7b4ac95ddda',
-          typeId: 'category',
-        },
-      ],
-      categoryOrderHints: {},
-      createdAt: '2024-05-23T06:06:16.344Z',
-      description: {
-        'en-US': 'A vibrant mix of seasonal spring flowers, including tulips, daisies, and lilies',
-      },
-      hasStagedChanges: false,
-      id: '85e1f1ea-d573-4c2a-881e-21529403f62d',
-      key: 'mixed_spring_bouquet',
-      lastModifiedAt: '2024-06-04T13:20:15.426Z',
-      masterVariant: {
-        attributes: [
-          {
-            name: 'color',
-            value: ['blue', 'pink', 'red', 'yellow', 'green'],
-          },
-          {
-            name: 'size',
-            value: 'medium',
-          },
-        ],
-        id: 1,
-        images: [
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/mixed_spring_bouquet-f3yDyp-m.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/mixed_spring_bouquet-1uadh3AI.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/mixed_spring_bouquet-sBT8dwqb.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/mixed_spring_bouquet-j1QKZOxN.jpg',
-          },
-        ],
-        prices: [
-          {
-            id: 'b8b48907-9bd3-4a29-9d6a-5af0f126e808',
-            key: 'mixed_spring_bouquet_price',
-            validFrom: '2024-05-30T21:00:00.000Z',
-            validUntil: '2024-06-29T21:00:00.000Z',
-            value: {
-              centAmount: 2599,
-              currencyCode: 'EUR',
-              fractionDigits: 2,
-              type: 'centPrecision',
-            },
-          },
-        ],
-      },
-      metaDescription: {},
-      metaTitle: {},
-      name: {
-        en: 'Mixed Spring Bouquet',
-        'en-US': 'Mixed Spring Bouquet',
-      },
-      priceMode: 'Embedded',
-      productType: {
-        id: '8ca55f27-b8ef-4321-aa32-7313517e00ba',
-        typeId: 'product-type',
-      },
-      published: true,
-      searchKeywords: {},
-      slug: {
-        'en-US': 'mixed-spring-bouquet',
-      },
-      variants: [],
-      version: 24,
-    },
-    {
-      categories: [
-        {
-          id: '3244f5bb-a30e-415e-8b43-faba85bb826e',
-          typeId: 'category',
-        },
-      ],
-      categoryOrderHints: {},
-      createdAt: '2024-05-23T06:27:33.187Z',
-      description: {
-        'en-US': 'A cascading bridal bouquet with white roses and delicate ivy.',
-      },
-      hasStagedChanges: false,
-      id: 'fdb627dd-acc2-4c01-9022-356829aeca90',
-      key: 'bridal_rose_cascade',
-      lastModifiedAt: '2024-06-04T13:20:15.187Z',
-      masterVariant: {
-        attributes: [
-          {
-            name: 'color',
-            value: ['white', 'green'],
-          },
-          {
-            name: 'size',
-            value: 'small',
-          },
-        ],
-        id: 1,
-        images: [
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/bridal_rose_cascade_-UBCTKH5m.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/bridal_flower_halo_s-8Lk2-zJe.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/bridal_rose_cascade_-xKgIQaKm.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/bridal_rose_cascade_-05X7jMw5.jpg',
-          },
-        ],
-        prices: [
-          {
-            discounted: {
-              discount: {
-                id: 'ff29af8a-b8fa-4e31-9b41-d6e7322c3aec',
-                typeId: 'product-discount',
-              },
-              value: {
-                centAmount: 15999,
-                currencyCode: 'EUR',
-                fractionDigits: 2,
-                type: 'centPrecision',
-              },
-            },
-            id: '82885822-b340-4af2-9bbb-b3c7ce5c4e7a',
-            key: 'bridal_rose_cascade_standard_price',
-            validFrom: '2024-05-30T21:00:00.000Z',
-            validUntil: '2024-06-29T21:00:00.000Z',
-            value: {
-              centAmount: 19999,
-              currencyCode: 'EUR',
-              fractionDigits: 2,
-              type: 'centPrecision',
-            },
-          },
-        ],
-      },
-      metaDescription: {},
-      metaTitle: {},
-      name: {
-        en: 'Bridal Rose Cascade',
-        'en-US': 'Bridal Rose Cascade',
-      },
-      priceMode: 'Embedded',
-      productType: {
-        id: 'd454363b-80d6-4736-827c-33051a33d045',
-        typeId: 'product-type',
-      },
-      published: true,
-      searchKeywords: {},
-      slug: {
-        'en-US': 'bridal-rose-cascade',
-      },
-      variants: [],
-      version: 28,
-    },
-    {
-      categories: [
-        {
-          id: '3244f5bb-a30e-415e-8b43-faba85bb826e',
-          typeId: 'category',
-        },
-      ],
-      categoryOrderHints: {},
-      createdAt: '2024-05-23T06:28:49.965Z',
-      description: {
-        'en-US': 'An elegant table centerpiece featuring lilies, roses, and hydrangeas.',
-      },
-      hasStagedChanges: false,
-      id: 'fbb04cf4-4dd7-4b7c-baa7-9950124a69c9',
-      key: 'elegant_centerpiece',
-      lastModifiedAt: '2024-06-04T13:20:15.353Z',
-      masterVariant: {
-        attributes: [
-          {
-            name: 'color',
-            value: ['blue', 'pink', 'white'],
-          },
-          {
-            name: 'size',
-            value: 'medium',
-          },
-        ],
-        id: 1,
-        images: [
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/elegant_centerpiece_-Ha-gFw7S.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/elegant_centerpiece_-Ip2vEWu3.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/elegant_centerpiece_-5F-337rd.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/elegant_centerpiece_-5b7akRjQ.jpg',
-          },
-        ],
-        prices: [
-          {
-            discounted: {
-              discount: {
-                id: 'ff29af8a-b8fa-4e31-9b41-d6e7322c3aec',
-                typeId: 'product-discount',
-              },
-              value: {
-                centAmount: 26399,
-                currencyCode: 'EUR',
-                fractionDigits: 2,
-                type: 'centPrecision',
-              },
-            },
-            id: 'c350ae1d-9504-4409-b523-bfbde3be3fc6',
-            key: 'elegant_centerpiece_standard_price',
-            validFrom: '2024-05-30T21:00:00.000Z',
-            validUntil: '2024-06-29T21:00:00.000Z',
-            value: {
-              centAmount: 32999,
-              currencyCode: 'EUR',
-              fractionDigits: 2,
-              type: 'centPrecision',
-            },
-          },
-        ],
-      },
-      metaDescription: {},
-      metaTitle: {},
-      name: {
-        en: 'Elegant Centerpiece',
-        'en-US': 'Elegant Centerpiece',
-      },
-      priceMode: 'Embedded',
-      productType: {
-        id: 'd454363b-80d6-4736-827c-33051a33d045',
-        typeId: 'product-type',
-      },
-      published: true,
-      searchKeywords: {},
-      slug: {
-        'en-US': 'elegant-centerpiece',
-      },
-      variants: [],
-      version: 26,
-    },
-    {
-      categories: [
-        {
-          id: '9204336a-dd71-4284-b1fd-358dc5b70de7',
-          typeId: 'category',
-        },
-      ],
-      categoryOrderHints: {},
-      createdAt: '2024-05-23T06:30:00.600Z',
-      description: {
-        'en-US': 'A serene arrangement of white lilies and greenery for a heartfelt tribute.',
-      },
-      hasStagedChanges: false,
-      id: 'd3f18f09-befc-4bf9-a6ce-198ebea730a2',
-      key: 'peaceful_white_lilies',
-      lastModifiedAt: '2024-06-04T13:20:15.432Z',
-      masterVariant: {
-        attributes: [
-          {
-            name: 'color',
-            value: ['white'],
-          },
-          {
-            name: 'size',
-            value: 'medium',
-          },
-        ],
-        id: 1,
-        images: [
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/peaceful_white_lilie-SaMrWT0f.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/peaceful_white_lilie-TbF1rl-_.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/peaceful_white_lilie-F44wv0ge.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/peaceful_white_lilie-4cmoUEkI.jpg',
-          },
-        ],
-        prices: [
-          {
-            id: '1ca2c7dd-c9b3-45a1-9da7-732b9f3ce410',
-            key: 'peaceful_white_lilies_standard_price',
-            validFrom: '2024-05-30T21:00:00.000Z',
-            validUntil: '2024-06-29T21:00:00.000Z',
-            value: {
-              centAmount: 1599,
-              currencyCode: 'EUR',
-              fractionDigits: 2,
-              type: 'centPrecision',
-            },
-          },
-        ],
-      },
-      metaDescription: {},
-      metaTitle: {},
-      name: {
-        en: 'Peaceful White Lilies',
-        'en-US': 'Peaceful White Lilies',
-      },
-      priceMode: 'Embedded',
-      productType: {
-        id: '472b9f4b-2028-4da6-9bd5-3e27f75a19f3',
-        typeId: 'product-type',
-      },
-      published: true,
-      searchKeywords: {},
-      slug: {
-        'en-US': 'peaceful-white-lilies',
-      },
-      variants: [],
-      version: 23,
-    },
-    {
-      categories: [
-        {
-          id: '9204336a-dd71-4284-b1fd-358dc5b70de7',
-          typeId: 'category',
-        },
-      ],
-      categoryOrderHints: {},
-      createdAt: '2024-05-23T06:30:49.392Z',
-      description: {
-        'en-US': 'A cross-shaped floral arrangement with roses and carnations.',
-      },
-      hasStagedChanges: false,
-      id: 'e1738d89-9540-4fab-8e96-331ee80e09d6',
-      key: 'comforting_cross',
-      lastModifiedAt: '2024-06-04T13:20:15.168Z',
-      masterVariant: {
-        attributes: [
-          {
-            name: 'color',
-            value: ['pink', 'green'],
-          },
-          {
-            name: 'size',
-            value: 'big',
-          },
-        ],
-        id: 1,
-        images: [
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/comforting_cross_sta-smo2svUh.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/comforting_cross_sta-q3wmRl2r.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/comforting_cross_sta-9W3B0s0n.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/comforting_cross_sta-3mM8fsXs.jpg',
-          },
-        ],
-        prices: [
-          {
-            id: '217adf35-abaf-4cf7-88c6-bd70da2d81cc',
-            key: 'comforting_cross_standard_price',
-            validFrom: '2024-05-30T21:00:00.000Z',
-            validUntil: '2024-06-29T21:00:00.000Z',
-            value: {
-              centAmount: 29999,
-              currencyCode: 'EUR',
-              fractionDigits: 2,
-              type: 'centPrecision',
-            },
-          },
-        ],
-      },
-      metaDescription: {},
-      metaTitle: {},
-      name: {
-        en: 'Comforting Cross',
-        'en-US': 'Comforting Cross',
-      },
-      priceMode: 'Embedded',
-      productType: {
-        id: '472b9f4b-2028-4da6-9bd5-3e27f75a19f3',
-        typeId: 'product-type',
-      },
-      published: true,
-      searchKeywords: {},
-      slug: {
-        'en-US': 'comforting-cross',
-      },
-      variants: [],
-      version: 23,
-    },
-    {
-      categories: [
-        {
-          id: 'd1592da9-553f-44e2-924d-2288d6566f47',
-          typeId: 'category',
-        },
-      ],
-      categoryOrderHints: {},
-      createdAt: '2024-05-23T06:31:28.786Z',
-      description: {
-        'en-US': 'A low-maintenance snake plant in a decorative pot, perfect for indoor spaces.',
-      },
-      hasStagedChanges: false,
-      id: 'c5c10a45-c4e6-4c42-a8c3-57f4c5215a6b',
-      key: 'snake_plant',
-      lastModifiedAt: '2024-06-04T13:20:15.444Z',
-      masterVariant: {
-        attributes: [
-          {
-            name: 'color',
-            value: ['green'],
-          },
-          {
-            name: 'size',
-            value: 'small',
-          },
-        ],
-        id: 1,
-        images: [
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/snake_plant_standard-VyOAw-L5.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/snake_plant_standard-ZNXsOES9.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/snake_plant_standard-_z_bhUnP.jpg',
-          },
-          {
-            dimensions: {
-              h: 1024,
-              w: 1024,
-            },
-            url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/snake_plant_standard-xokoo0jC.jpg',
-          },
-        ],
-        prices: [
-          {
-            id: '350cbbc6-70b2-4ef6-8b4b-072e70e7c449',
-            key: 'snake_plant_standard_price',
-            validFrom: '2024-05-30T21:00:00.000Z',
-            validUntil: '2024-06-29T21:00:00.000Z',
-            value: {
-              centAmount: 1299,
-              currencyCode: 'EUR',
-              fractionDigits: 2,
-              type: 'centPrecision',
-            },
-          },
-        ],
-      },
-      metaDescription: {},
-      metaTitle: {},
-      name: {
-        en: 'Snake Plant',
-        'en-US': 'Snake Plant',
-      },
-      priceMode: 'Embedded',
-      productType: {
-        id: '79744707-8ee7-452e-ac68-dde481eba9da',
-        typeId: 'product-type',
-      },
-      published: true,
-      searchKeywords: {},
-      slug: {
-        'en-US': 'snake-plant',
-      },
-      variants: [],
-      version: 23,
-    },
-  ],
-  total: 30,
-};
-
-const results = [
-  {
-    categories: [
-      {
-        id: 'b86a46f7-70a9-483f-b74b-a7b4ac95ddda',
-        typeId: 'category',
-      },
-    ],
-    categoryOrderHints: {},
-    createdAt: '2024-05-23T05:54:46.722Z',
-    description: {
-      'en-US': 'A timeless bouquet featuring a dozen fresh red roses, elegantly arranged with greenery.',
-    },
-    hasStagedChanges: false,
-    id: '5a2b875e-68f9-42c2-83f7-67f5418a55ca',
-    key: 'classic_rose_bouquet',
-    lastModifiedAt: '2024-06-04T13:20:15.329Z',
-    masterVariant: {
-      attributes: [
-        {
-          name: 'color',
-          value: ['green', 'red'],
-        },
-        {
-          name: 'size',
-          value: 'medium',
-        },
-      ],
-      id: 1,
-      images: [
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/classic_rose_bouquet-oBl_4aEu.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/classic_rose_bouquet-gBL67E9D.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/classic_rose_bouquet-SVfPmFL7.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/classic_rose_bouquet-N496xh6R.jpg',
-        },
-      ],
-      prices: [
-        {
-          id: '1df0bacc-adf1-4d3f-9dbd-8882def9f079',
-          key: 'classic_rose_bouquet_standard_price',
-          validFrom: '2024-05-30T21:00:00.000Z',
-          validUntil: '2024-06-29T21:00:00.000Z',
-          value: {
-            centAmount: 2999,
-            currencyCode: 'EUR',
-            fractionDigits: 2,
-            type: 'centPrecision',
-          },
-        },
-      ],
-    },
-    metaDescription: {},
-    metaTitle: {},
-    name: {
-      en: 'Classic Rose Bouquet',
-      'en-US': 'Classic Rose Bouquet',
-    },
-    priceMode: 'Embedded',
-    productType: {
-      id: '8ca55f27-b8ef-4321-aa32-7313517e00ba',
-      typeId: 'product-type',
-    },
-    published: true,
-    searchKeywords: {},
-    slug: {
-      'en-US': 'classic-rose-bouquet',
-    },
-    variants: [],
-    version: 25,
-  },
-  {
-    categories: [
-      {
-        id: 'b86a46f7-70a9-483f-b74b-a7b4ac95ddda',
-        typeId: 'category',
-      },
-    ],
-    categoryOrderHints: {},
-    createdAt: '2024-05-23T06:06:16.344Z',
-    description: {
-      'en-US': 'A vibrant mix of seasonal spring flowers, including tulips, daisies, and lilies',
-    },
-    hasStagedChanges: false,
-    id: '85e1f1ea-d573-4c2a-881e-21529403f62d',
-    key: 'mixed_spring_bouquet',
-    lastModifiedAt: '2024-06-04T13:20:15.426Z',
-    masterVariant: {
-      attributes: [
-        {
-          name: 'color',
-          value: ['blue', 'pink', 'red', 'yellow', 'green'],
-        },
-        {
-          name: 'size',
-          value: 'medium',
-        },
-      ],
-      id: 1,
-      images: [
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/mixed_spring_bouquet-f3yDyp-m.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/mixed_spring_bouquet-1uadh3AI.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/mixed_spring_bouquet-sBT8dwqb.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/mixed_spring_bouquet-j1QKZOxN.jpg',
-        },
-      ],
-      prices: [
-        {
-          id: 'b8b48907-9bd3-4a29-9d6a-5af0f126e808',
-          key: 'mixed_spring_bouquet_price',
-          validFrom: '2024-05-30T21:00:00.000Z',
-          validUntil: '2024-06-29T21:00:00.000Z',
-          value: {
-            centAmount: 2599,
-            currencyCode: 'EUR',
-            fractionDigits: 2,
-            type: 'centPrecision',
-          },
-        },
-      ],
-    },
-    metaDescription: {},
-    metaTitle: {},
-    name: {
-      en: 'Mixed Spring Bouquet',
-      'en-US': 'Mixed Spring Bouquet',
-    },
-    priceMode: 'Embedded',
-    productType: {
-      id: '8ca55f27-b8ef-4321-aa32-7313517e00ba',
-      typeId: 'product-type',
-    },
-    published: true,
-    searchKeywords: {},
-    slug: {
-      'en-US': 'mixed-spring-bouquet',
-    },
-    variants: [],
-    version: 24,
-  },
-  {
-    categories: [
-      {
-        id: '3244f5bb-a30e-415e-8b43-faba85bb826e',
-        typeId: 'category',
-      },
-    ],
-    categoryOrderHints: {},
-    createdAt: '2024-05-23T06:27:33.187Z',
-    description: {
-      'en-US': 'A cascading bridal bouquet with white roses and delicate ivy.',
-    },
-    hasStagedChanges: false,
-    id: 'fdb627dd-acc2-4c01-9022-356829aeca90',
-    key: 'bridal_rose_cascade',
-    lastModifiedAt: '2024-06-04T13:20:15.187Z',
-    masterVariant: {
-      attributes: [
-        {
-          name: 'color',
-          value: ['white', 'green'],
-        },
-        {
-          name: 'size',
-          value: 'small',
-        },
-      ],
-      id: 1,
-      images: [
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/bridal_rose_cascade_-UBCTKH5m.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/bridal_flower_halo_s-8Lk2-zJe.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/bridal_rose_cascade_-xKgIQaKm.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/bridal_rose_cascade_-05X7jMw5.jpg',
-        },
-      ],
-      prices: [
-        {
-          discounted: {
-            discount: {
-              id: 'ff29af8a-b8fa-4e31-9b41-d6e7322c3aec',
-              typeId: 'product-discount',
-            },
-            value: {
-              centAmount: 15999,
-              currencyCode: 'EUR',
-              fractionDigits: 2,
-              type: 'centPrecision',
-            },
-          },
-          id: '82885822-b340-4af2-9bbb-b3c7ce5c4e7a',
-          key: 'bridal_rose_cascade_standard_price',
-          validFrom: '2024-05-30T21:00:00.000Z',
-          validUntil: '2024-06-29T21:00:00.000Z',
-          value: {
-            centAmount: 19999,
-            currencyCode: 'EUR',
-            fractionDigits: 2,
-            type: 'centPrecision',
-          },
-        },
-      ],
-    },
-    metaDescription: {},
-    metaTitle: {},
-    name: {
-      en: 'Bridal Rose Cascade',
-      'en-US': 'Bridal Rose Cascade',
-    },
-    priceMode: 'Embedded',
-    productType: {
-      id: 'd454363b-80d6-4736-827c-33051a33d045',
-      typeId: 'product-type',
-    },
-    published: true,
-    searchKeywords: {},
-    slug: {
-      'en-US': 'bridal-rose-cascade',
-    },
-    variants: [],
-    version: 28,
-  },
-  {
-    categories: [
-      {
-        id: '3244f5bb-a30e-415e-8b43-faba85bb826e',
-        typeId: 'category',
-      },
-    ],
-    categoryOrderHints: {},
-    createdAt: '2024-05-23T06:28:49.965Z',
-    description: {
-      'en-US': 'An elegant table centerpiece featuring lilies, roses, and hydrangeas.',
-    },
-    hasStagedChanges: false,
-    id: 'fbb04cf4-4dd7-4b7c-baa7-9950124a69c9',
-    key: 'elegant_centerpiece',
-    lastModifiedAt: '2024-06-04T13:20:15.353Z',
-    masterVariant: {
-      attributes: [
-        {
-          name: 'color',
-          value: ['blue', 'pink', 'white'],
-        },
-        {
-          name: 'size',
-          value: 'medium',
-        },
-      ],
-      id: 1,
-      images: [
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/elegant_centerpiece_-Ha-gFw7S.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/elegant_centerpiece_-Ip2vEWu3.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/elegant_centerpiece_-5F-337rd.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/elegant_centerpiece_-5b7akRjQ.jpg',
-        },
-      ],
-      prices: [
-        {
-          discounted: {
-            discount: {
-              id: 'ff29af8a-b8fa-4e31-9b41-d6e7322c3aec',
-              typeId: 'product-discount',
-            },
-            value: {
-              centAmount: 26399,
-              currencyCode: 'EUR',
-              fractionDigits: 2,
-              type: 'centPrecision',
-            },
-          },
-          id: 'c350ae1d-9504-4409-b523-bfbde3be3fc6',
-          key: 'elegant_centerpiece_standard_price',
-          validFrom: '2024-05-30T21:00:00.000Z',
-          validUntil: '2024-06-29T21:00:00.000Z',
-          value: {
-            centAmount: 32999,
-            currencyCode: 'EUR',
-            fractionDigits: 2,
-            type: 'centPrecision',
-          },
-        },
-      ],
-    },
-    metaDescription: {},
-    metaTitle: {},
-    name: {
-      en: 'Elegant Centerpiece',
-      'en-US': 'Elegant Centerpiece',
-    },
-    priceMode: 'Embedded',
-    productType: {
-      id: 'd454363b-80d6-4736-827c-33051a33d045',
-      typeId: 'product-type',
-    },
-    published: true,
-    searchKeywords: {},
-    slug: {
-      'en-US': 'elegant-centerpiece',
-    },
-    variants: [],
-    version: 26,
-  },
-  {
-    categories: [
-      {
-        id: '9204336a-dd71-4284-b1fd-358dc5b70de7',
-        typeId: 'category',
-      },
-    ],
-    categoryOrderHints: {},
-    createdAt: '2024-05-23T06:30:00.600Z',
-    description: {
-      'en-US': 'A serene arrangement of white lilies and greenery for a heartfelt tribute.',
-    },
-    hasStagedChanges: false,
-    id: 'd3f18f09-befc-4bf9-a6ce-198ebea730a2',
-    key: 'peaceful_white_lilies',
-    lastModifiedAt: '2024-06-04T13:20:15.432Z',
-    masterVariant: {
-      attributes: [
-        {
-          name: 'color',
-          value: ['white'],
-        },
-        {
-          name: 'size',
-          value: 'medium',
-        },
-      ],
-      id: 1,
-      images: [
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/peaceful_white_lilie-SaMrWT0f.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/peaceful_white_lilie-TbF1rl-_.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/peaceful_white_lilie-F44wv0ge.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/peaceful_white_lilie-4cmoUEkI.jpg',
-        },
-      ],
-      prices: [
-        {
-          id: '1ca2c7dd-c9b3-45a1-9da7-732b9f3ce410',
-          key: 'peaceful_white_lilies_standard_price',
-          validFrom: '2024-05-30T21:00:00.000Z',
-          validUntil: '2024-06-29T21:00:00.000Z',
-          value: {
-            centAmount: 1599,
-            currencyCode: 'EUR',
-            fractionDigits: 2,
-            type: 'centPrecision',
-          },
-        },
-      ],
-    },
-    metaDescription: {},
-    metaTitle: {},
-    name: {
-      en: 'Peaceful White Lilies',
-      'en-US': 'Peaceful White Lilies',
-    },
-    priceMode: 'Embedded',
-    productType: {
-      id: '472b9f4b-2028-4da6-9bd5-3e27f75a19f3',
-      typeId: 'product-type',
-    },
-    published: true,
-    searchKeywords: {},
-    slug: {
-      'en-US': 'peaceful-white-lilies',
-    },
-    variants: [],
-    version: 23,
-  },
-  {
-    categories: [
-      {
-        id: '9204336a-dd71-4284-b1fd-358dc5b70de7',
-        typeId: 'category',
-      },
-    ],
-    categoryOrderHints: {},
-    createdAt: '2024-05-23T06:30:49.392Z',
-    description: {
-      'en-US': 'A cross-shaped floral arrangement with roses and carnations.',
-    },
-    hasStagedChanges: false,
-    id: 'e1738d89-9540-4fab-8e96-331ee80e09d6',
-    key: 'comforting_cross',
-    lastModifiedAt: '2024-06-04T13:20:15.168Z',
-    masterVariant: {
-      attributes: [
-        {
-          name: 'color',
-          value: ['pink', 'green'],
-        },
-        {
-          name: 'size',
-          value: 'big',
-        },
-      ],
-      id: 1,
-      images: [
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/comforting_cross_sta-smo2svUh.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/comforting_cross_sta-q3wmRl2r.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/comforting_cross_sta-9W3B0s0n.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/comforting_cross_sta-3mM8fsXs.jpg',
-        },
-      ],
-      prices: [
-        {
-          id: '217adf35-abaf-4cf7-88c6-bd70da2d81cc',
-          key: 'comforting_cross_standard_price',
-          validFrom: '2024-05-30T21:00:00.000Z',
-          validUntil: '2024-06-29T21:00:00.000Z',
-          value: {
-            centAmount: 29999,
-            currencyCode: 'EUR',
-            fractionDigits: 2,
-            type: 'centPrecision',
-          },
-        },
-      ],
-    },
-    metaDescription: {},
-    metaTitle: {},
-    name: {
-      en: 'Comforting Cross',
-      'en-US': 'Comforting Cross',
-    },
-    priceMode: 'Embedded',
-    productType: {
-      id: '472b9f4b-2028-4da6-9bd5-3e27f75a19f3',
-      typeId: 'product-type',
-    },
-    published: true,
-    searchKeywords: {},
-    slug: {
-      'en-US': 'comforting-cross',
-    },
-    variants: [],
-    version: 23,
-  },
-  {
-    categories: [
-      {
-        id: 'd1592da9-553f-44e2-924d-2288d6566f47',
-        typeId: 'category',
-      },
-    ],
-    categoryOrderHints: {},
-    createdAt: '2024-05-23T06:31:28.786Z',
-    description: {
-      'en-US': 'A low-maintenance snake plant in a decorative pot, perfect for indoor spaces.',
-    },
-    hasStagedChanges: false,
-    id: 'c5c10a45-c4e6-4c42-a8c3-57f4c5215a6b',
-    key: 'snake_plant',
-    lastModifiedAt: '2024-06-04T13:20:15.444Z',
-    masterVariant: {
-      attributes: [
-        {
-          name: 'color',
-          value: ['green'],
-        },
-        {
-          name: 'size',
-          value: 'small',
-        },
-      ],
-      id: 1,
-      images: [
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/snake_plant_standard-VyOAw-L5.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/snake_plant_standard-ZNXsOES9.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/snake_plant_standard-_z_bhUnP.jpg',
-        },
-        {
-          dimensions: {
-            h: 1024,
-            w: 1024,
-          },
-          url: 'https://images.cdn.us-central1.gcp.commercetools.com/0e9a8cb4-e1d4-4736-bb77-1646edd8e799/snake_plant_standard-xokoo0jC.jpg',
-        },
-      ],
-      prices: [
-        {
-          id: '350cbbc6-70b2-4ef6-8b4b-072e70e7c449',
-          key: 'snake_plant_standard_price',
-          validFrom: '2024-05-30T21:00:00.000Z',
-          validUntil: '2024-06-29T21:00:00.000Z',
-          value: {
-            centAmount: 1299,
-            currencyCode: 'EUR',
-            fractionDigits: 2,
-            type: 'centPrecision',
-          },
-        },
-      ],
-    },
-    metaDescription: {},
-    metaTitle: {},
-    name: {
-      en: 'Snake Plant',
-      'en-US': 'Snake Plant',
-    },
-    priceMode: 'Embedded',
-    productType: {
-      id: '79744707-8ee7-452e-ac68-dde481eba9da',
-      typeId: 'product-type',
-    },
-    published: true,
-    searchKeywords: {},
-    slug: {
-      'en-US': 'snake-plant',
-    },
-    variants: [],
-    version: 23,
-  },
-];
-
-export { data, results };
